factor out and disable gsm secrets
