+++ ecoscope_workflows/_settings/_gsm.py
@@ -111,22 +111,3 @@ class GoogleSecretManagerConfigSettingsSource(PydanticBaseSettingsSource):
             logger.debug(e)
 
         return d
-
-
-class _Settings(BaseSettings):
-    @classmethod
-    def settings_customise_sources(
-        cls,
-        settings_cls: type[BaseSettings],
-        init_settings: PydanticBaseSettingsSource,
-        env_settings: PydanticBaseSettingsSource,
-        dotenv_settings: PydanticBaseSettingsSource,
-        file_secret_settings: PydanticBaseSettingsSource,
-    ) -> tuple[PydanticBaseSettingsSource, ...]:
-        return (
-            init_settings,
-            env_settings,
-            dotenv_settings,
-            file_secret_settings,
-            GoogleSecretManagerConfigSettingsSource(settings_cls),
-        )
