+++ ecoscope_workflows/_settings/_settings.py
@@ -1,0 +1,20 @@
+from pydantic_settings import BaseSettings, PydanticBaseSettingsSource
+
+
+class _Settings(BaseSettings):
+    @classmethod
+    def settings_customise_sources(
+        cls,
+        settings_cls: type[BaseSettings],
+        init_settings: PydanticBaseSettingsSource,
+        env_settings: PydanticBaseSettingsSource,
+        dotenv_settings: PydanticBaseSettingsSource,
+        file_secret_settings: PydanticBaseSettingsSource,
+    ) -> tuple[PydanticBaseSettingsSource, ...]:
+        return (
+            init_settings,
+            env_settings,
+            # dotenv_settings,
+            # file_secret_settings,
+            # GoogleSecretManagerConfigSettingsSource(settings_cls),
+        )
